moving extract to its own file
needed both on windows / unix platforms
